+++ packages/jmix-react-antd/src/util/history.ts
@@ -1,4 +1,4 @@
-import {addPagingParams} from "../ui/paging/Paging";
+import {addPagingParams} from "@haulmont/jmix-react-web";
 import {JmixPagination} from "@haulmont/jmix-react-core";
 
 export function saveHistory(routingPath: string, pagination?: JmixPagination) {
