+++ packages/jmix-react-web/src/index.ts
@@ -22,8 +22,10 @@ export * from "./ui/MultiScreen";
 export * from "./ui/MultilineText";
 export * from "./ui/Msg";
 export * from "./ui/Tabs/Content";
-export * from "./ui/paging/pagingConfig";
+export * from "./ui/paging/config";
+export * from "./ui/paging/utils";
 export * from "./ui/entity-property/EntityProperty";
+export * from './ui/form/FieldPermssionContainer';
 
 export * from './screen-registration/screen-registration';
