+++ packages/jmix-react-web/src/ui/paging/config.ts
@@ -1,0 +1,22 @@
+/**
+ * Paging defaults
+ */
+ export const defaultPaginationConfig = {
+    current: 1,
+    pageSize: 10,
+    defaultPageSize: 10,
+    showSizeChanger: true,
+    pageSizeOptions: ['10', '20', '50', '100', '500', '1000', '5000']
+  };
+  
+  export const defaultGridPaginationConfig = {
+    current: 1,
+    pageSize: 12,
+    defaultPageSize: 12,
+    showSizeChanger: true,
+    pageSizeOptions: ['12', '24', '48', '96', '240', '960', '4800']
+  };
+  
+  /**
+   * Ant Design Pagination component adopted to our List and DataCollectionStore
+   */
