+++ packages/jmix-react-web/src/ui/paging/utils.test.ts
@@ -1,0 +1,91 @@
+import { addPagingParams, parsePagingParams, setPagination, createPagingConfig, PaginationConfigBase } from './utils'
+import { defaultPaginationConfig } from './config'
+
+describe('addPagingParams', () => {
+  it('should return same url if one of param not set', () => {
+    expect(addPagingParams('initialUrl', undefined, undefined)).toBe('initialUrl');
+    expect(addPagingParams('initialUrl', undefined, 3)).toBe('initialUrl');
+    expect(addPagingParams('initialUrl', 2, undefined)).toBe('initialUrl');
+  });
+
+  it('should add paging params to url', () => {
+    expect(addPagingParams('initialUrl', 2, 4)).toBe('initialUrl?page=2&pageSize=4');
+  });
+});
+
+describe('parsePagingParams', () => {
+  it('should parse null paging params', () => {
+    expect(parsePagingParams('initialUrl', undefined, undefined))
+      .toEqual({});
+    expect(parsePagingParams('', undefined, undefined))
+      .toEqual({});
+  });
+
+  it('should parse paging params', () => {
+    expect(parsePagingParams('?page=4&pageSize=7', undefined, undefined))
+      .toEqual({current: 4, pageSize: 7});
+    expect(parsePagingParams('?page=3', 10, 20))
+      .toEqual({current: 3, pageSize: 20});
+    expect(parsePagingParams('?pageSize=8', 10, 20))
+      .toEqual({current: 10, pageSize: 8});
+
+    expect(parsePagingParams('?page=&pageSize=', 7, 18))
+      .toEqual({current: 7, pageSize: 18});
+    expect(parsePagingParams('?page=AAAAA&pageSize=BBBBBB', 7, 18))
+      .toEqual({current: 7, pageSize: 18});
+    expect(parsePagingParams('?page=null&pageSize=undefined', 7, 18))
+      .toEqual({current: 7, pageSize: 18});
+  });
+
+  it ('should parse paging params without question mark at start', () => {
+    expect(parsePagingParams('page=4&pageSize=7', undefined, undefined))
+      .toEqual({current: 4, pageSize: 7});
+  })
+});
+
+  const pagingConfig: PaginationConfigBase = {
+    current: 5,
+    pageSize: 10
+  };
+
+describe('setPagination', () => {
+
+  it('should set pagination to dataCollection', () => {
+      let ds = {offset: 0, limit: 0, load: jest.fn()} as any;
+      setPagination(pagingConfig, ds);
+      expect(ds).toMatchObject({offset: 40, limit: 10});
+      expect(ds.load).not.toBeCalled();
+
+      ds = {offset: 0, limit: 0, load: jest.fn()} as any;
+      setPagination({}, ds);
+      expect(ds).toMatchObject({offset: 0, limit: 0});
+      expect(ds.load).not.toBeCalled();
+  });
+
+});
+
+describe('createPagingConfig', () => {
+
+  it('should create disabled paging config', () => {
+    expect(createPagingConfig('', true).disabled).toBeTruthy();
+  });
+
+  it('should create enabled paging config', () => {
+    expect(createPagingConfig(''))
+      .toMatchObject(defaultPaginationConfig);
+  });
+
+  it('should check that page size is used only if it\'s value in pageSizeOptions', () => {
+    expect(createPagingConfig('?page=4&pageSize=7'))
+      .toMatchObject(defaultPaginationConfig);
+
+    expect(createPagingConfig('?page=4&pageSize=10'))
+      .toMatchObject({...defaultPaginationConfig, current: 4, pageSize: 10});
+
+    const config = {...defaultPaginationConfig, pageSizeOptions: ['7']};
+    expect(createPagingConfig('?page=4&pageSize=7', false, config))
+      .toMatchObject({...config, current: 4, pageSize: 7});
+  });
+
+});
+
